Removed TreeBuilderActions and moved its impl into TreeBuilder
